# openapi.yaml
openapi: 3.0.1
info:
  title: Canteen Management API
  description: API documentation for the Canteen Modern Web App (Zomato/Swiggy Clone)
  version: v1.0
  contact:
    name: Canteen Support
    email: support@canteen.com
servers:
- url: http://localhost:8080
  description: Local development server
...
paths:
  /api/menu:
    get:
      tags:
      - Menu Management
      summary: Get all menu items
      description: Retrieves a list of all available food items in the canteen
      operationId: getAllMenuItems
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MenuItem'
    post:
      tags:
      - Menu Management
      summary: Add a new menu item
      description: Creates a new food item in the menu
      operationId: createMenuItem
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MenuItem'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MenuItem'
...
components:
  schemas:
    MenuItem:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        imageUrl:
          type: string
        category:
          type: string
        available:
          type: boolean
        rating:
          type: number
          format: double
        preparationTime:
          type: integer
          format: int32
...